Overworld field moves do not require knowledge
In overworld events/interactions, all pokémon that can learn a field
move are able to perform that field move.
Basically the "CheckPartMove" function has been modified to return true
if a mon can learn the requested TM/HM (requested from an overworld
interaction), instead of if it knows it.
Logic has also been added so that the respective TM/HM has to be in the
bag for the mon to perform the move.