Menu field moves do not require knowledge
Pokémon can use field moves outside of battle as long as they can learn
them. Show the allowed field moves in the menu.
As a compromise, only the moves that are not mainly triggered by an
overworld interaction/event (i.e., fly, flash, dig) are added to the
list. Teleport and health snack are not considered, to make their usage
more demanding/interesting.
Also added a failsafe so that no more than 4 moves are attempted to be
shown at once (the game will crash otherwise). Learned moves have
priority. The others are prioritized in the previously stated order:
fly > flash > dig.
TODO: Only add items if the player has the respective HM/TM in the bag.
TODO: Remove fly from the list if not learned and cannot be used(?)